ADD: pico, lploc, camvas libraries in components and utils directory from static to src
